audio-chunker: send serialised events before and after each chunk
That way we can still tell where chunk boundaries are after
the audio has gone through an encoder which will output
many small chunks.

Note that the chunk-end event might not make it through an
audio encoder immediately after a chunk in all cases, it might
only be pushed when the next chunk comes into the encoder.