srt-fragment-enc: make flac output work consistently
FLAC frames contain a frame number (or sample offset number) which
starts from 0 and counts up. We need to overwrite that with a number
based on the absolute timestamp in order to make sure that we get
bit-identical frames for the same timestamp no matter when we
started the receiver or whether there are discontinuities in the
audio. And once we've done that we also need to fix up a header
crc8 checksum and the frame crc16 checksum at the end.