aes67-relay: use udpsink sync=true with audiotestsrc
We switched audiotestsrc to is-live=false in order to get
consistent/predictable timestamps, but that means that
we should use sync=true on the output's sink element,
otherwise there's no throttling and we're just flooding
the output.

For live inputs like udpsrc/sdp and srtsrc we set the
sinks to sync=false because there's not much point in
properly syncing to the clock in our particular case,
as the data flow is steady in any case.