fragment-enc: discard chunks after discontinuity based on number of frames
Don't just blindly discard the first 10 chunks at the beginning or
after a discontinuity, without taking into account the number of
frames per chunk.

Instead, define the stabilisation requirement in number of frames
and keep chunks once we've seen enough continuously coded audio
frames, assuming the encoder has stabilised the bitstream by then.

This usually means that in most configurations it's probably enough
to just discard the first chunk at the beginning or after a discont,
instead of the first 10.