audiochunker: use metas to delineate chunk start/end boundaries
... instead of custom events, to bring down latency.

The problem was that the custom event sent after the last buffer
of the chunk would only make it through the audio encoder once
there's more data to push out, so only once the next chunk is
ready, adding at least one (if not more, on packet loss) chunks
latency.

With metas the chunk end meta makes it through the audio encoder
immediately on the last encoded audio frame.